Move analyzer into Carter project
