+++ Carter.sln
@@ -32,10 +32,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Carter.ResponseNegotiators.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Carter.ResponseNegotiators.Newtonsoft.Tests", "test\Carter.ResponseNegotiators.Newtonsoft.Tests\Carter.ResponseNegotiators.Newtonsoft.Tests.csproj", "{47DD3AA8-F072-41E3-AF7F-119DBBD8D14A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Carter.Analyzers", "src\Carter.Analyzers\Carter.Analyzers.csproj", "{14D66E34-E351-441D-B71D-443ED69AC43C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Carter.Analyzers.Tests", "test\Carter.Analyzers.Tests\Carter.Analyzers.Tests.csproj", "{5B978E31-83AB-4632-A39F-1C87B5F49A30}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -154,30 +150,6 @@ Global
 		{47DD3AA8-F072-41E3-AF7F-119DBBD8D14A}.Release|x64.Build.0 = Release|Any CPU
 		{47DD3AA8-F072-41E3-AF7F-119DBBD8D14A}.Release|x86.ActiveCfg = Release|Any CPU
 		{47DD3AA8-F072-41E3-AF7F-119DBBD8D14A}.Release|x86.Build.0 = Release|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Debug|x64.Build.0 = Debug|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Debug|x86.Build.0 = Debug|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Release|x64.ActiveCfg = Release|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Release|x64.Build.0 = Release|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Release|x86.ActiveCfg = Release|Any CPU
-		{14D66E34-E351-441D-B71D-443ED69AC43C}.Release|x86.Build.0 = Release|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Debug|x64.Build.0 = Debug|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Debug|x86.Build.0 = Debug|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Release|x64.ActiveCfg = Release|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Release|x64.Build.0 = Release|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Release|x86.ActiveCfg = Release|Any CPU
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -192,8 +164,6 @@ Global
 		{E5D99119-96EB-4C01-9EC2-A06FC3EEC182} = {35DE35A0-758D-4FDD-BDA3-67F04F65677D}
 		{D30E87B0-39AE-41FE-8BA3-E5F9FBFFDD64} = {5D056A8A-821C-4C3C-A281-4FA7F8CE251B}
 		{47DD3AA8-F072-41E3-AF7F-119DBBD8D14A} = {DCB5B9A0-F06D-4BDF-917D-A459C790C718}
-		{14D66E34-E351-441D-B71D-443ED69AC43C} = {5D056A8A-821C-4C3C-A281-4FA7F8CE251B}
-		{5B978E31-83AB-4632-A39F-1C87B5F49A30} = {DCB5B9A0-F06D-4BDF-917D-A459C790C718}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9096DE78-6327-48BA-AE0E-336F769681A7}
