+++ src/Carter/DependencyContextAssemblyCatalog.cs
@@ -89,7 +89,9 @@ public class DependencyContextAssemblyCatalog
     {
         try
         {
+#pragma warning disable RS1035
             return Assembly.Load(assemblyName);
+#pragma warning restore RS1035
         }
         catch (Exception)
         {
